refactor: Statistics 패키지 이동
Statistics를 util에서 domain으로 이동
